requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
azure-identity>=1.15.0
azure-ai-openai>=1.0.0
pypdf>=4.0.0
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.1.0
pydantic>=2.5.0
typing-extensions>=4.8.0
python-magic>=0.4.27
graphviz>=0.20.1 